move: 연광성 있는 객체끼리 뭉치게 변경
