+++ src/main/java/nextstep/ladder/domain/ladder/Ladder.java
@@ -1,5 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
 
+import nextstep.ladder.domain.player.Players;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
