+++ src/main/java/nextstep/ladder/domain/ladder/Position.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
 
 import java.util.Objects;
