+++ src/main/java/nextstep/ladder/domain/ladder/RandomPointStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
